implement zip
